requirements.txt
asgiref==3.4.1
attrs==21.2.0
coverage==6.2
django==3.2.10
djangorestframework==3.13.0
importlib-metadata==4.8.2
iniconfig==1.1.1
jsonformatter==0.3.1
packaging==21.3
parameterized==0.8.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
uwsgi==2.0.20
zipp==3.6.0
